bart/requirements.txt
requests
py-rouge
nltk